+++ src/app/view-config-upgraders.js
@@ -345,3 +345,16 @@ export function upgradeFrom1_0_9(config) {
     version: '1.0.10',
   };
 }
+
+// TODO
+
+// Added in version 1.0.14:
+// - Support for the optional 'coordinationValues' property for views.
+export function upgradeFrom1_0_13(config) {
+  const newConfig = cloneDeep(config);
+
+  return {
+    ...newConfig,
+    version: '1.0.14',
+  };
+}